chore: replace eventually crate with v0.5.0-alpha.1 version
