--- eventually-v0-5-0/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "eventually-v0-5-0"
-version = "0.5.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-async-trait = "0.1.51"
-futures = "0.3.17"
-serde = { version = "1.0.130", features = ["derive"] }
-thiserror = "1.0.30"
-
-[dev-dependencies]
-anyhow = "1.0.51"  # NOTE: this is only used for test components and assertions.
-lazy_static = "1.4.0"
-serde_json = "1.0.69"
-tokio = { version = "1.13.0", features = ["macros", "rt-multi-thread"] }
