--- eventually/src/util.rs
@@ -1,8 +1,0 @@
-//! Collection of utilities that extends or implements some of the traits
-//! found in this crate.
-
-pub mod inmemory;
-pub mod optional;
-pub mod sync;
-
-pub use tokio::spawn;
