--- eventually/src/util/sync.rs
@@ -1,4 +1,0 @@
-//! Module containing the asynchronous synchronization primitives
-//! the crate uses in the public API.
-
-pub use tokio::sync::RwLock;
